mozreview: copy commit extra_data to draft_extra_data on draft creation (Bug 1244448). r=mdoglio r=mcote r=gps
In order to mimic the semantics of the built-in extra data for
ReviewRequest and ReviewRequestDraft we need to copy
CommitData.extra_data to CommitData.draft_extra_data when a draft is
created.

Handling clearing the data isn't necessary as we will always gate access
to draft_extra_data by checking if the draft exists first. Any stale
data will be overwritten when a draft is next created for the
ReviewRequest.