reviewboard: store commit ID in commit messages instead of extras (bug 1243486); r=dminor
Storing the commit ID in extras was littered with problems. First, the
ID wasn't always available from data that Review Board had access to (it
sees the commit message but not the extras by default). Second,
Mercurial doesn't always preserve extras, as the existing `hg graft`
test demonstrated. While there wasn't a test for it, I believe MQ
dropped the custom extras entry as well. Third, Git stores the commit
ID in the commit message. Both implementations should be consistent to
make logic using the commit IDs simpler.

This commit moves the commit ID storage from extras to the commit
message. Because we changed the content of the commit, SHA-1s change.
This resulted in massive test fallout. Most of the changes were basic
overwrites of the new SHA-1. Some test commands referenced SHA-1s and
these needed updated. In a few cases, uses of `hg commit --amend -m`
had to define the commit ID manually to carry it over so a new one
wasn't generated on next use of the changeset.

The old code for installing a hacked up "rebase" command that preserves
extras has been removed.

MozReview-Commit-ID: 1WyO65ojxuv