reviewboard: add test for review request recycling (bug 1242246); r=dminor
The bug report indicated a failure to map a commit when obsolescence was
enabled and a commit is deleted and a new one added. Our current
behavior is to recycle leftover review requests, which is wrong if we
know for sure the incoming commit is not logically related to the old
one.

We add a test demonstrating the buggy behavior.

MozReview-Commit-ID: LcZTmoxt6j9