mozreview: don't reuse review requests if commit is new (bug 1242246); r=dminor
If we now an incoming commit is not logically related to an unclaimed
one, we should create a new review request instead of recycling a review
request that belongs to an unrelated commit.

test-specify-reviewers.t changed slightly because it has obsolescence
enabled and exercises a code path where a new commit is pushed to an
existing series.

MozReview-Commit-ID: LRi48ckMsWq