testing: remove 'activity-indicator' check from autoland selenium tests
It appears due to recent changes the 'activity-indicator' no longer appears.
In test-autoland-inbound.py and test-autoland-try.py it was used to check
for reviewboard server errors. Removing this check means that it will be
necessary to call self.dump_reviewboard_log() in the test to identify if a
server error is the reason the test has failed. I think this is an
acceptable cost to pay for having a test that runs.

Unfortunately, test-autoland-concurrent.py relies upon this check for its
basic functionality. I've filed Bug 1246950 for this and am marking the
test as skipped until it is fixed.