ansible/hg-web: remove CRON to normalize hgweb permissions
The CRON consumes a lot of I/O and generates Kafka/ZooKeeper alerts
around 2130 PST every night due to presumably the cluster not being able
to process requests in time due to high I/O load.

With the recent change to set the umask of the vcsreplicator processors
to 002, the file permissions should always be proper and this cron
should not be necessary.