mozhg: make prompt default argument an empty string (bug 1247044); r=smacleod
Mercurial 3.7 enforces that the value for "default" be a string or
character buffer object. We were previously passing None, which is
neither. This was causing the prompt to throw a TypeError on Mercurial
3.7.

Since we're only checking truthiness of the result, it doesn't matter if
it is an empty string or None, since they are both Falsy. So use an
empty string.

MozReview-Commit-ID: 9O9FPbnokMu