hgmo: web command for exposing automation relevance (bug 1247802); r=dminor
I had planned to expose JSON info for automation relevance via
Mercurial's built-in JSON API. Unfortunately, the search functionality
of that API is not yet implemented. Bug 1247322 is on file to track
that. I looked into implementing it, but it would be a few times more
work than implementing a trivial one-off web command that exposes the
same data.

This commit introduces a web command for exposing the results of an
automationrelevant() revset query for a revision. It calls into the
hgweb functions for populating changelist entries, which means our
extensions are called to populate custom metadata. So we get pushlog
info in the output. Yay.

MozReview-Commit-ID: BzXUyAoCJQo