hgmo: prune some keys from the returned JSON (bug 1248850); r=dminor
This is a very hacky way to avoid a templating issue due to changelogtag
missing. Fixing the underlying issue requires handling special return
values from Mercurial's internal templating API and I have little
desire to do that because those APIs are fragile. It is easier to just
remove the output that calls into them to avoid the problem.

MozReview-Commit-ID: BmcB5LnfJkj