hgmo: serve AWS clients bundles from local S3 region (bug 1249197); r=fubar, dminor
This commit modifies the hgmo extension to inspect the source IP
address of HTTP requests for bundle clone and clone bundles manifests
and compare against known AWS IP ranges. If the request is coming from
an AWS region that we have S3 URLs for (identified by the presence of
"ec2region=" in the manifest metadata), we filter the manifest to only
advertise URLs in the same region. This guarantees an intra-region
transfer without any client configuration. This is fast and free
since intra-region S3 transfers don't cost anything!

In addition, we also reorder the manifest to advertise the stream clone
bundles first. We recommend clients in automation apply stream clone
bundles (over the default gzip bundles) because they are the fastest
mechanism to clone repositories (a stream clone bundle can apply at
>80 MB/s).

Tests for the new features have been added to the hgserver tests because
we have existing tests that run the code for generating manifests
and we want to use realistic values for testing.

Documentation for the bundle hosting has been updated to reflect the
change in behavior.

MozReview-Commit-ID: 2ZaoJpaiqBU