hgserver: pash command to associate LDAP account with MozReview (bug 1249149); r=dminor
In terms of security and compatibility with existing workflows, the easiest way to
associate a LDAP account with MozReview is over SSH. This is because an
SSH connection proves ownership of that LDAP account and SSH is already
how we associate LDAP accounts today.

This commit introduces a pash SSH command for associating an LDAP
account with MozReview. It prompts for a Bugzilla username and API Key
then makes a HTTP request to Review Board's LDAP association web API.

Everything behaves just like LDAP association is currently done except
we're not doing it from a Mercurial context.

We introduce a standalone script for performing the HTTP request because
Pash is a Python 2.6 process running out of the system Python
installation and therefore doesn't have access to the Review Board
Python package. We run the script out of the Python 2.7 virtualenv,
which is where pash-invoked `hg` processes run out of.

Comprehensive tests of the new functionality have been added.

Documentation has also been added.

We've talked about performing a reverse LDAP search to automatically
associate LDAP accounts. However, this may not always work (e.g. not
all users may have Bugzilla metadata registered in LDAP). So, a manual
solution (which this commit provides) will likely always need to exist.
While the manual association in this commit isn't the most turn-key,
it does unblock HTTP pushers (including all Git users) from not being
able to use several MozReview features. Hopefully that will be enough
until the LDAP search functionality is deployed (if ever).

MozReview-Commit-ID: 6OYua0RyeA8