mozreview: Backout 7014000ff305 for making it harder to copy/paste (bug 1249297).
This was originally landed as part of a feature in bug 1101601, but there has
been significant opposition to this feature since landing (mainly because it
is either annoying to some people defaulted on, and undiscoverable if defaulted
off).

We're aiming for a more general solution that makes the comment action more
obvious. See bug 1249647.

MozReview-Commit-ID: BeTC93mcIG8