hghooks: allow pushing to try without a try syntax (bug 1247072), r=gps
Try syntax is no longer required to schedule jobs on try. This can now be accomplished
directly in treeherder (or via other mozci based tools). This changes the try hook to
allow pushes with no try syntax. For now, print a verbose WARNING so people understand
what's happening. We may want to tone the warning down a bit if this workflow becomes
more standard in the future.

MozReview-Commit-ID: AKgkBpSxHdd