ansible: securely install pip 8.0.3 and setuptools 20.1.1 in CentOS Docker image
Previously we were running older versions of both. We're trying to
upgrade to pip 8 everywhere since it has hash verification built-in.

As part of switching to pip 8, we add a verifying URL downloader to
protect against tampering.

We also switch URLs to our S3 bucket, which is under our control and
should be more reliable.

MozReview-Commit-ID: 3HfiLCF9qWN