ansible/hg: use pip 8 (bug 1243501); r=fubar
pip 8 contains peep's hash verification functionality.

As part of this commit, we introduce a standalone task for configuring a
virtualenv. We replace the virtualenv muckery with it. And we
update the requirements.txt files to pip 8's syntax.

After this patch, only Autoland is using peep. The reason it wasn't
converted is because it is using its own mechanism and wasn't easily
ported like everything else. It can be done later.

There is a non-zero chance that this will break a prod deploy
of MozReview and/or hg.mo. We should be cautious the first time
we deploy.

fubar reviewed the first version of this commit. It had to be
reworked a bit to incorporate other changes that were made
since, such as the addition of Pygments to some virtualenvs.

MozReview-Commit-ID: F70nbqBd4xK