ansible/mozreview: install Python 2.7
The MozReview web heads are still running Python 2.6 from the system
Python installation. We want to run Python 2.7 everywhere.

The first step towards transitioning rbweb to Python 2.7 is to get
Python 2.7 installed. We do this by establishing a new "rbweb" role
and have it install Python 2.7. We add this role in various locations.

It's worth noting that the admin and web machines are currently more
under the control of Puppet than Ansible. We'll need to slowly bring
more things under Ansible control in order to switch to actually running
things from Python 2.7.