testing: upgrade run-tests.py
Imported run-tests.py from changeset c7f89ad87bae from upstream. This
corresponds to a version partially through the 3.8 release cycle.

This gets us a number of bug fixes.

Our fixes to increase port allocation and perform DOCKER_HOSTNAME
substitution have been added back in. I should really spend time to make
these configurable upstream so we don't need to modify run-tests.py...

We had to revert a change to _hghave() regarding locating the hghave
binary. Upstream changed it to look for hghave next to run-tests.py.
We could refactor our tree to put our custom hghave in the same
directory. This is probably for the better. For now, it is scope bloat.