bundleclone: make testing compatible with Mercurial 3.7 (bug 1228088); r=dminor
The bundleclone extension no longer works in Mercurial 3.7 because
localrepository.clone (the function it overrides to do most of the
special clone work) no longer exists.

The extension no-ops when the server supports clone bundles anyway. So
it was already doing nothing on 3.6 when clone bundles was enabled.

We update the tests to not run on 3.7+. We also add a new test that
verifies 3.7+ continues to work (as a no-op).

MozReview-Commit-ID: HTp3aOJ4MUB