mozext: use modern revset APIs (bug 1228088); r=dminor
Mercurial 3.7 is issuing warnings when running tests that the revset
functions in mozext are returning lists instead of revset classes.

mozext was one of the first Mercurial extensions I wrote. It is one of
the oldest and doesn't exactly contain the best patterns or modern API
usage.

This commit brings the revset code into the modern age. Most of the
changes involve calling subset.filter() to lazily evaluate a revset.
Before, we iterated over the subset, which meant that we had to evaluate
every element before returning. This almost certainly made a number of
revset queries slower than they should have been.

MozReview-Commit-ID: 6mmWIXgfEAo