mozreview: add reviewer delegation (bug 1161230) r=smacleod
Adds the ability for anyone with the required permissions in BMO to update the
target reviewers.  Due to limitations Review Board's limited permissions model
(only the submitter can make changes), this requires a localstorage hosted
draft, updating Review Board as the submitter, and providing a sudo-like
ability for updating Bugzilla.

MozReview-Commit-ID: H51WMF4SCTx