generate-hg-s3-bundles: parse values from fields and uniform parsing (bug 1257371); r=dminor
Before, we assumed fields after the repo name were simple sets of flags.
Now, we allow "key=value" strings to be defined. This will be used in a
subsequent commit to specify the source repository that a manifest
should be copied from.

While we're here, we unify parsing so command arguments and lines from a
file are parsed the same way. This will make testing easier.

MozReview-Commit-ID: 8LRFrhNDO1Z