ansible/hg-ssh: use mozilla-central's bundles for the try repo (bug 1257371); r=dminor
Currently, the try repository doesn't advertise bundles. The Try
repository is effectively a superset of mozilla-central, aurora, beta,
release, esr, and tons of other repos. We don't yet have a unified repo
from which to generate a bundle. But a bundle is better than no bundle.
And the bundle with the most changesets is mozilla-central. So let's
advertise that on the Try repo.

MozReview-Commit-ID: JzonWV0SWSX