ansible: create /var/hg (bug 1260604); r=fubar
We currently use /repo/hg for holding various Mercurial related files.
First, this is unusal because it doesn't conform with expectations for
normal UNIX filesystem layout.

Second, /repo/hg had its origins in the days where it was a NFS mount
everywhere. And, on hgssh it still is an NFS mount. This is causing
problems in production because Ansible races to modify files on a
shared filesystem. Surprisingly, this hasn't bit us too much. However,
imminent changes to roll out a new OS to hgssh will press the issue.
A shared filesystem with e.g. different Python binaries running on
different machines will completely bork the virtualenvs. So, let's
prepare to start moving things to /var/hg and do away with most
of the foo in /repo/hg.

MozReview-Commit-ID: V7MVfKVdEE