Bug 1261127 - Copy scripts from hgscripts repo; r=fubar
The added files are copies of scripts from changeset f6918f5ce0cb
of https://hg.mozilla.org/hgcustom/hgscripts. Some of these scripts
may not be used any more. I'd rather add them verbatim now and clean up
things later than possibly forget to add scripts to the repo.

The change is in preparation of wiping the content of the hgscripts
repo. After this commit, that repo is fully redundant with this one.