scripts: add a standalone script for performing SSH key lookup from LDAP (bug 1261212); r=fubar, kang
Modern versions of OpenSSH support calling a program that will print SSH
public keys for a user. This script added by this patch is intended to be
used for this purpose.

We go beyond simple key lookup by filtering out DSA keys because they
are insecure and shouldn't be allowed. We stop short of filtering out
RSA keys with a short modulus, however. I don't want to introduce a
BER decoder just yet (we would likely import paramiko's).

MozReview-Commit-ID: 2bYNOsaF76C