ansible: switch hgmaster to CentOS 7 (bug 1261212); r=fubar, kang
As part of the switch, the sshd config has changed significantly.
Before, we used openssh with the LPK (LDAP) patches. It listened on
port 22 and processed root and hg logins.

In the new world, we have 2 sshd daemons. The system sshd listens on
port 22 and behaves like a normal sshd. The hg sshd listens on port 222.
It disallows root login (yay!). LDAP integration is handled by the
AuthorizedKeysCommand calling the custom LDAP key lookup script
we just implemented.

The new hg sshd configuration is using modern and recommended OpenSSH
settings. We now generate and prefer ED25519 keys. DSA keys are gone.
The RSA key length has increased from 2048 to 4096 bits.

CentOS 7 also brings systemd to the table. The new sshd for hg is
managed via systemd. Kafka and Zookeeper are still managed via
supervisor. We will likely want to convert these to systemd someday.
But it's easier to leave them as supervisor for now since they are still
running on CentOS 6 for hgweb.

On the Docker side of things, the hg-ssh role now needs to start more
system services via supervisor. I couldn't get systemd working in Docker
and the internet tells me it is pain, pain, and more pain. For some
reason, system LDAP integration didn't work in CentOS 7 until I started
nscd in the Docker container. I'm not sure why. But having nscd running
is a good thing because that's how production works.

Since we're no longer using openssh-lpk, a lot of the code from
the openssh-lpk Ansible role has been copied into the hg-ssh role. This
could probably be split into a new role. I imagine we'll do that when we
need to convert reviewboard-hg to CentOS 7. For now, I'm fine with the
duplication. If nothing else, it minimizes the risk of CentOS 6 and
CentOS 7 interfering with each other in the Ansible configs.

MozReview-Commit-ID: HIDMv539Apm