mozreview: stop clearing bugzilla flags on comments (bug 1197879); r=mcote,smacleod
Now that the reviewer can explicitly set review flags, we can stop clearing
those flags on comment. As a result, a flag set on mozreview is always mirrored to bugzilla.
In case of a r? -> r+ -> r? status transition, the final flag will have the reviewer set as
both the setter and the requestee.

MozReview-Commit-ID: Ao2OiaAOsSd