ansible/hg-ssh: configure repo unification daemon (bug 1108729)
We add config files to to define the unification that occurs.

A systemd service to run the unification daemon is added.

We don't yet enable the systemd service on any machines. We'll do this
manually.

We likely also need to add some syslog configuration so the logs
for the daemon process are logged into a standalone file. Let's see what
happens first...

MozReview-Commit-ID: 4AV1pmjUbdI