mozreview: associate ldap accounts in start-local-mozreview (bug 1265972) r=smacleod
Users created with |mozreview create-user| are not automatically
associated with their ldap account, so we make the association as
part of the script. This is more convenient than requiring a
developer to setup the association manually, since it is most likely
desired.

However, it is important to keep these operations separate, and have
association be optional, since this is how production behaves and we
use it for testing.

MozReview-Commit-ID: JXJVGrp21MB