testing: start Pulse container as part of hg cluster (bug 1196915); r?smacleod
We're about to add Pulse integration to hg.mozilla.org. Before we can do
this, we need the Docker testing cluster to run a Pulse server so we can
test the new features.

Fortunately, we've already implemented a Pulse Docker container as part
of the MozReview testing environment. So, we re-use that image. "All" we
need to do is hook up the Pulse container to the hg cluster. This is a
lot of code changes, but it's pretty boilerplate.

MozReview-Commit-ID: JC8fs2Tci1y