testing: add more hashes for feedparser
Before, we only captured the tar.bz2 hash. Something on PyPI must
have changed because it now doesn't pull down the tar.bz2 file by
default, leading to a hash mismatch. We add the hashes for the
tar.gz and zip files so make sure whatever it downloads is verified.