testing: add more Pillow package hashes
Again, PyPI seems to have changed its behavior and is now advertising
different archive types first or something. So, make sure we list both
the zip and tar.gz source archives so we don't get a hash verification
error.

The packages file for the MozReview virtualenv does not list the
hash of the OS X wheel because that environment is Linux, hence why
it has 2 instead of 3 hashes.