ansible: add a generic mechanism to email systemd unit output
As we move CRONs to systemd, we need a mechanism to alert people
when a CRON has failed. There isn't a built-in mechanism in systemd
to do this. What the Internet suggests is you create a generic
service unit which executes something that sends a mail. You then
use OnFailure in your original unit file to invoke the generic
failure service.

This commit implements such a system. Our generic service invokes
a Python script which invokes journalctl to read the last few minutes
of data for a given unit. It then sends an email with the result.

MozReview-Commit-ID: 5muzykWkjEg