ansible/hg-ssh: replace bundle generation CRON with a systemd timer unit
Systemd can do periodic tasks. It works better than CRON because it will
prevent services from running multiple times in parallel and it
integrates with the journal, so we will always have timestamped
output of CRONs.

The service uses our mail-unit-output generic service to send emails on
failure. We install a drop-in file to specify where output should go.

MozReview-Commit-ID: 7VFq7fZ8Zta