ansible/hg-ssh: create hg-master.target systemd unit (bug 1268131); r=fubar
There are multiple services that need to run on the hg master server and
only the active hg master server. We create a systemd target unit to
control them as a group.

The target has a condition on a file being present on the NFS mount that
specifies the current active master. This should prevent the target from
starting unless it is the current master server.

MozReview-Commit-ID: ED76GGWqhVi