ansible/hg-web: unset push_ssl to get rid of "ssl required" message during HTTP push (bug 949153)
Mercurial only allows pushing via https:// by default because security.
When you push to a http:// server, it prints a "ssl required" error
message.

The problem with our load balancer setup is the load balancer terminates
TLS and passes plain text to the origin servers. So Mercurial thinks it
is plain text and prints the "ssl required" message even though the
client is using https://.

We unset push_ssl in the config. The side effect is Mercurial now prints
an "authorization required" message because by default Mercurial
disallows pushing.

This error message is still kind of crappy. But it is less misleading
than "ssl required."