ansible/hg-web: finish converting vcsreplicator to a systemd service (bug 1268722)
After testing the last commit in production, we're ready to move forward
with the conversion of the vcsreplicator service from supervisor to
systemd.

We added Ansible configs to enable the 8 instances of the service (1 for
each partition).

The old supervisor config file has been removed. The supervisor
dependency has been removed from hg-web.

Because we don't run systemd in the Docker environment, we preserved the
supervisor config in the docker-hg-web role. The supervisor role has
been added as a dependency.