firefoxtree: support serving firefoxtree tags from bookmarks (bug 1264814); r=ahal
The experimental unified Firefox repo advertises bookmarks for the
various Firefox repo heads. Instead of teaching that repo to have a
firefoxtreetags file, let's teach it to serve the special firefoxtree
tags from the defined bookmarks instead.

This patch also changes the client pull code to filter bookmarks from
the remote that match firefoxtree tags, but only if the remote is
advertising firefoxtree tags.

MozReview-Commit-ID: FBiKR2iGo1n