robustcheckout: add extension to perform clone+checkout robustly (bug 1273305); r=smacleod
Firefox automation has multiple implementations of code that ensures
a Mercurial repository is checked out at a specified path. None of
these implementations are optimal and there are numerous efficiency
gains to be realized.

This commit introduces the "robustcheckout" extension. It provides
a command (`hg robustcheckout`) which effectively performs a
clone/pull + purge + update robustly using optimal techniques for
automation. These include the required use of pooled shared stores.

The goal is to replace the custom code in automation with this
extension, starting with hgtool.

The code in this commit comes from the mozharness work I did in
bug 1270317. It also incorporates the content of the "purgelong"
extension, which was previously a standalone extension only used
as part of doing a clone/pull+checkout in automation.

MozReview-Commit-ID: 9szgfyqtnHM