mozreview: don't reset r- flags when reviewers are added (bug 1273790); r=smacleod
When a review request is updated we need to handle prior review states
correctly.  If the update contains changes to the diff, we need to carry
forward r+'s, and reset all other flags to r?.  If the update does not update
the diff then all flags shouldn't be touched.

MozReview-Commit-ID: CRm7HGTtKWO