robustcheckout: reject --revision arguments that aren't SHA-1s (bug 1274095); r=jlund
We take shortcuts with regards to handling the --revision argument: if
the value resolves to a known revision, we skip pulling from the remote
and checkout directly. This is faster. But it assumes --revision is
constant. --revision isn't constant if it is symbolic. e.g. "default"
or "tip." To prevent these kinds of errors, reject values to --revision
that don't resemble SHA-1 fragments.

We enforce the check in 2 locations: first as a visual check during
argument processing then again when checking the revision is known
locally. The 2nd only occurs if the revision's value looks like a SHA-1
fragment but isn't a known SHA-1 fragment.

MozReview-Commit-ID: 98O2F7ye6mD