firefoxtree: remove "review" path (bug 1275122); r=smacleod
This was added in the very early days of MozReview. I just watched
it confuse a new contributor because they couldn't figure out where the
"review" path was coming from.

This automagical definition of this path is nice. But if anything it
should be part of the mozreview extension, not firefoxtree.

MozReview-Commit-ID: Cbhe1xX7cCx