configwizard: add support for setting optimal diff settings (bug 1277406); r=glob
This is almost a straight port of the legacy wizard. We dropped setting
diff.unified=8 because this is holdover from the days when we
exported patches to Bugzilla after `hg export`. These days we have
tools to submit to MozReview and attach to Bugzilla. These tools should
provide appropriate context.

But diff.git and diff.showfunc are generally useful, so keep them
enabled.

MozReview-Commit-ID: FMJhkEMaWNk