configwizard: prompt for username; implement config saving (bug 1277406); r=glob
This commit does a lot.

The next part of the config wizard to port is setting the username.
Since Mercurial's built-in config API doesn't support writing config
files, we need to use a separate API for that. We borrow the approach
of the old config wizard and use configobj for config file
manipulation. However, we change things up a bit.

Instead of using ConfigObj for reading and writing, we use it for
writing only. This ensures that config values from %include are
recognized by the wizard. We also remove the restriction of the
old config parser that prohibited %include from working. We do this
by normalizing %include during load and save operations.

MozReview-Commit-ID: DGBg2eQpqdF