reviewboard: don't use "&>" in tests
"&>" is a bash 4+ operator. In other shells, it backgrounds a task.
This was causing race conditions in tests and tests to fail pretty
reliably.

Replace with "2>&1" which is a more portable way of doing this.

Before this commit, I would see various failures running tests with -j1.
After this commit, all tests pass. I suppose the backgrounded tasks were
still running after test termination and would interfering with
subsequent tests. Ouch.