configwizard: prompt to enable and configure pager (bug 1279596); r=glob
Lack of pager is IMO one of the most annoying things in a default
Mercurial config.

Enabling the pager extension gets a partially useful pager. But
I feel you have to configure it a bit more, namely enabling it
on other commands and defining a custom `less` invocation to
make it behave better. `less` without -F in particular is quite
annoying because you have to exit the pager for every command,
even if it prints a single line of output!

The config added in this commit mostly matches my local config.
I think it is a good starting point.

MozReview-Commit-ID: 4WTStWybmfr