vcsreplicator: attempt to fix race condition in test-check-aggregator-lag.t
There is a race between the just-written messages being consumed by
the aggregator and the check running. Wait for consumers to ack the
messages and add a sleep to give the aggregator time to process them.

The sleep is a bit hacky. But the aggregator polls every 100ms or
so, so hopefully it is fine.