configwizard: handle missing cmdutil.command (bug 1282293); r=glob
cmdutil.command was introduced in hg 1.9. Working around it being
missing allows the extension to load on hg 1.7 and 1.8.

It might run on earlier versions. However, I'm unable to run hg
1.6 and earlier on my machine when building hg from source. I
think compatibility down to 1.7 should be fine. So unless someone
complains, I'm inclined to leave this as our cut-off.

MozReview-Commit-ID: 1Vpy5V10Tj6