hgmo: decode description and reuse variable
ctx.description() is a mercurial.encoding.localstr. This is a str
like type that is encoded to Mercurial's active encoding (which
is likely utf-8 in our server environment). By using
encoding.fromlocal() we get the raw bytes of the commit message.
This is what we should be using in case there are any weird
character sequences that get normalized by local encoding.

In addition, ctx.description() has some overhead associated with
it. So we save the description into a local variable to avoid
having to do redundant work.

MozReview-Commit-ID: LKrHSJWo64h