mozreview: add functions for retrieving diffstat information (Bug 1286017). r=glob
This commit ignores extra information Review Board has, such as
information on how many lines were replaced, rather than just deleted
and inserted. This is done to keep things simple as the extra
information isn't always present and I haven't thought of a good way
to conditionally pass things along.

This also makes consuming this new data much simpler as two numbers
which will always be present are used. These numbers are also pretty
self explanatory.

In the future we might want to expand this, but we can always keep
the current numbers we provide and just add new information.

MozReview-Commit-ID: JJSiQH51qm5