pushlog: query against unfiltered repo (bug 1286426); r=glandium
By querying against an unfiltered repo, we are able to find hidden
changesets referenced by the pushlog.

Test changes demonstrate that queries against hidden changesets now
yield values.

We still have failures for "full=1" queries because we get a
filtering error when attempting to retrieve the changectx for a
hidden changeset. And, hidden changesets are still being referenced in
JSON because we don't check for hidden state in the presentation layer.
This will be addressed in subsequent commits.

MozReview-Commit-ID: 3i23L9katSB