pushlog: filter hidden changesets (bug 1286426); r=glandium
If a changeset in the pushlog results is filtered, we detect this and
discard the changeset.

Future commits will continue to refine this behavior. For now, we gain
test coverage to make it easier to see how things change (or don't change)
going forward.

The docs for the ``changeset`` key have been tweaked.

MozReview-Commit-ID: KtkEhIdFfQ3