pushlog: expose obsolescence metadata (bug 1286426); r=glandium
The next step towards handling obsolescence data in pushlog is to expose
it to consumers so they can do intelligent things.

This commit changes behavior in a number of ways.

First, "successors" and "precursors" nodes lists are added to
changeset entries if the repository has obsolescence data.

Second, a "obsoletechangesets" key is now present on each push entry.
It operates like "changesets" except it only exposes changesets
that are obsolete.

I decided to create a new container for obsolete changesets because
adding obsolete changesets to the existing container is too risky:
existing consumers of the pushlog wouldn't know to e.g. look for
a key to identify a changeset as obsolete. By introducing a new
container, existing consumers have to explicitly opt in to looking
at obsolete changesets. And, since key additions are backwards
compatible, we don't need to introduce a v3 of the API to expose
this new data!

There should be enough metadata in the pushlog now for consumers to
determine whether a changeset was rewritten and what the
previous or new changeset is. We may also want to update Pulse
notifications to include similar metadata so Pulse consumers don't
need to consult the pushlog.

MozReview-Commit-ID: 1H1Kwo7KSxc