pushlog: add tests for pulling entries with obsolete changesets (bug 1286426); r=glandium
The test demonstrates sub-optimal behavior when pulling normally. This is
because the remote server operates against the "visible" filter, which
doesn't expose obsoleted changesets.

We have a hack in the replication extension used on hg.mozilla.org to
allow replication access to the unfiltered repository on the server,
thus ensuring obsoleted changesets are transferred to clients during
regular `hg pull` operations. Tests for this will be added in a
different commit.

We also took the opportunity to update the docs about what pushlog
pulling does and when it can fail.

MozReview-Commit-ID: ERInCX0yLRY