obsolescencehacks: extension to enable obsolescence exchange under conditions (bug 1286022); r=glandium
Obsolescence markers are officially an experimental Mercurial feature.
One of the reasons they are experimental is that the exchange of
markers doesn't yet scale well, adding significant latency to
e.g. `hg pull`.

We want to enable obsolescence markers on some Firefox repositories
because they are a useful feature. e.g. when a user pushes to MozReview
and that changeset is rebased and eventually lands on central, clients
with obsolescence enabled can follow the rewrites and automatically
hide the rebased changesets or help users recover from rebases. But
due to scaling concerns, we aren't yet ready to accept obsolescence
markers from all users.

As a first step towards enabling obsolescence markers, we're going
to enable the creation of markers on the autoland repo (an upcoming
commit will create them server-side). However, we aren't yet ready
to enable the exchange of those markers via `hg push`. But, if
markers are being produced and marker exchange isn't enabled,
replication won't work properly. That's where this commit comes in.

The extension added in this commit provides a mechanism to enable
the obsolescence exchange feature - but only for a set of
config-defined users.

We update our infrastructure configuration to allow the user
performing replication to exchange markers. A side-effect
of this is that any repository with obsolescence enabled will
have marker exchange enabled via https://hg.mozilla.org/. However,
obsolescence exchange won't be enabled for ssh://hg.mozilla.org/
because exchange is only globally allowed for vcs-sync@mozilla.com
(the user replication mirrors use).

We initially plan to only enable markers for the autoland repo.
And only Mercurial clients that have enabled obsolescence markers
will fetch them. And, most users shouldn't fetch the autoland repo.
So hopefully there is a minimal performance impact to users and
we won't run into marker exchange scaling issues.

When we enable obsolescence markers for other repos (notably
mozilla-central), we will definitely want to evaluate whether
exposing markers via https://hg.mozilla.org/ is the correct
thing to do. For now, I think we can live with enabling for the
autoland repo.

Tests for basic extension functionality and functionality in the
hg.mozilla.org environment have been added and seem to exhibit
acceptable behavior.

MozReview-Commit-ID: I62ELNB69D4