Bug 1287234 - Do not merge stdout and stderr in get_output. r=gps
The commands get_output is run for may emit messages on stderr that are
only noise for the git mozreview script, altering the expected output.

With current git cinnabar tip of the release branch, a warning may be
emitted when running `git cinnabar hg2git`, which prevents `git mozreview
configure` from finding the right repository because the output contains
more than the expected sha1s.

MozReview-Commit-ID: 5HUvprewVP1